% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{plot_gamma_performance}
\alias{plot_gamma_performance}
\title{Plot Individual Gamma Estimates vs. Choice Slope}
\usage{
plot_gamma_performance(fit, data)
}
\arguments{
\item{fit}{A cmdstanr fit object.}

\item{data}{The prepared data frame used for fitting.}
}
\description{
Plot Individual Gamma Estimates vs. Choice Slope
}
