% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{plot_hierarchical_params}
\alias{plot_hierarchical_params}
\title{Visualize Group-Level Parameter Posteriors}
\usage{
plot_hierarchical_params(fit)
}
\arguments{
\item{fit}{A cmdstanr fit object.}
}
\description{
Visualize Group-Level Parameter Posteriors
}
