% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{plot_rt_curves}
\alias{plot_rt_curves}
\title{Plot RT Chronometric Curves}
\usage{
plot_rt_curves(data, observed_data = NULL, bin_gaze = FALSE)
}
\arguments{
\item{data}{Prepared data frame (usually predictions).}

\item{observed_data}{Optional. The raw data frame used for fitting.}

\item{bin_gaze}{Logical. If TRUE, splits curves by gaze advantage.}
}
\description{
Plot RT Chronometric Curves
}
