% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{plot_rt_fit}
\alias{plot_rt_fit}
\title{Plot Binned RT Fit Comparison (N-Item Race)}
\usage{
plot_rt_fit(observed, predicted, n_bins = 15)
}
\arguments{
\item{observed}{Prepared data frame of actual trials.}

\item{predicted}{Data frame of simulations from \code{predict_glam}.}

\item{n_bins}{Number of bins for the RT distribution.}
}
\description{
Plot Binned RT Fit Comparison (N-Item Race)
}
