% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glam.R
\name{fit_glam}
\alias{fit_glam}
\title{Fit N-Item GLAM via Hierarchical Bayesian Inference}
\usage{
fit_glam(
  data,
  method = "meanfield",
  s_val = NULL,
  force_recompile = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame prepared via \code{prep_glam_data}.}

\item{method}{Character, "meanfield" or "sample".}

\item{s_val}{Numeric. Fixed sensitivity. If NULL, estimated as free parameter.}

\item{force_recompile}{Logical. Defaults to FALSE.}

\item{...}{Additional arguments for cmdstanr.}
}
\description{
Fit N-Item GLAM via Hierarchical Bayesian Inference
}
