% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{plot_choice_curves}
\alias{plot_choice_curves}
\title{Plot Choice Psychometric Curves}
\usage{
plot_choice_curves(data, bin_gaze = FALSE)
}
\arguments{
\item{data}{Prepared data frame (observed or predicted).}

\item{bin_gaze}{Logical. If TRUE, splits curves by gaze advantage (Left vs Right).}
}
\description{
Plot Choice Psychometric Curves
}
