% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_glam.R
\name{predict_glam}
\alias{predict_glam}
\title{Generate Predictions from a Fitted GLAM}
\usage{
predict_glam(fit, newdata, n_sims = 10, type = "mean")
}
\arguments{
\item{fit}{A cmdstanr fit object.}

\item{newdata}{A data frame prepared via prep_glam_data.}

\item{n_sims}{Number of simulations per trial.}

\item{type}{Character, "mean" to use posterior means or "posterior" to sample.}
}
\description{
Generate Predictions from a Fitted GLAM
}
